CASMPET-4712: Expand range for allowed users/groups
We've recently changed the cray-service chart to default to setting
the user and group to run as to 65534 (the nobody user on the K8s
NCNs). While we've decided that this user is OK to run as the
Gatekeeper policy would still reject this since it's outside the
default range.

The range is now set to 1 - 65535. 0 is not in the allowed range
because that UID has root access.