CASMPET-4716: Clean up excludedNamespaces
The `loftsman` namespace is removed from excludedNamespaces. There's
nothing in it so there's no need to exclude it.

The `istio` namespace is removed from excludedNamespaces. We don't
create an `istio` namespace. I can't think of a reason that the 2
istio-related namespaces that we do have should be excluded from
Gatekeeper checks.