CASMTRIAGE-2849: gatekeeper hook use curl
The gatekeeper chart was failing to deploy properly because the
hook image configured to use kubectl but it calls curl. The fix
is to change the hook to use an image with curl in it.

The curl image used is chosen to comply with the
"Referencing common images" section in CASM-2670.